Fix: terminal stopping because large images
So i found a trouble with large images to set a color when run this script, the trouble it's caused because of the echos who can tell the deletion of the temporary file, and i made a update to get if the cava it's currently running or not in the system.

+++ shcava.sh
@@ -30,8 +30,6 @@ if [ -e "$input_file" ]; then
     echo "gradient_color_$((i+1)) = '$color'" >> "$output_file"
   done
 
-  echo "Conversão concluída. O resultado foi salvo em '$output_file'."
-  echo "O arquivo cores foi apagado!"
   rm "$input_file"
 else
   echo "Arquivo de entrada '$input_file' não encontrado."
@@ -39,5 +37,8 @@ fi
 
 # move as configurações atuais para a pasta Cava
 mv config $HOME/.config/cava/
-killall -SIGUSR1 cava
 
+if pgrep -x "cava" > /dev/null
+then
+    killall -SIGUSR1 cava
+fi
